# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "High-performance RESTful backend API for the Timesheet Automation Portal, built with Go (Gin Engine) and PostgreSQL.\n- Automated Timesheet Generator: Multi-company Excel (.xlsx) and PDF generation (MII, SDD, Adidata, NTT) with day-fraction calculations, dynamic day trimming, formula preservation, and async SMTP delivery.\n- Dynamic Template Engine: Dedicated programmatic template builders alongside visual cell-mapping coordinate parsers integrated with Handsontable.\n- Dual Authentication \u0026 Security: JWT sessions with NIST SP 800-63B compliant Argon2id password hashing and WebAuthn FIDO2 Passkeys (biometric \u0026 hardware-bound).\n- Role-Based Access Control (RBAC): Admin and User role segregation, centralized account administration, and profile change approval workflows.\n- Web Push Notifications: Daily reminder scheduler pinned to 17:00 WIB (Asia/Jakarta) notifying active users via VAPID.\n- 3NF Relational Master Data: Normalized PostgreSQL database schema (Companies, Departments, Projects, Holidays, Activity Statuses) with foreign keys and automated SQL migrations.\n- API Versioning: Standardized RESTful endpoints routed under the `/api/v1/*` namespace.",
        "title": "Timesheet Automation Portal API",
        "termsOfService": "https://github.com/naufalzaid17/timesheet-generator",
        "contact": {
            "name": "API Support"
        },
        "license": {
            "name": "MIT"
...
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/v1/activities": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieves all daily activities for the authenticated user for a specific month and year.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Activity"
                ],
                "summary": "List monthly activities",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Year (defaults to current year)",
                        "name": "year",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Month 1-12 (defaults to current month)",
                        "name": "month",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.DailyActivity"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Creates or updates a daily activity record for the authenticated user on a given date.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Activity"
                ],
                "summary": "Upsert daily timesheet activity",
                "parameters": [
                    {
                        "description": "Daily activity payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.DailyActivityRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.DailyActivity"
                        }
                    },
                    "400": {
                        "description": "Invalid date format or payload",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/activity-statuses": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns all normalized activity status options (e.g. Present, Sick, Vacation).",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Master Data"
                ],
                "summary": "List activity status options",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.ActivityStatus"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/api/v1/companies": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns all companies with their associated projects, templates, and departments.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Master Data"
                ],
                "summary": "List all companies",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Company"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/departments": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns active departments, optionally filtered by company_id.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Master Data"
                ],
                "summary": "List active departments",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Company ID filter",
                        "name": "company_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Department"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/holidays": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns public holidays for the specified month and year with database cache fallback.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Holiday"
                ],
                "summary": "Get monthly Indonesian public holidays",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Year (defaults to current year)",
                        "name": "year",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Month 1-12 (defaults to current month)",
                        "name": "month",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.HolidayDTO"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/holidays/all": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns holidays, optionally filtered by year or company_id.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Holiday"
                ],
                "summary": "List holidays",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Company ID filter",
                        "name": "company_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Year filter",
                        "name": "year",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Holiday"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.DailyActivity": {
            "type": "object",
            "properties": {
                "activity": {
                    "type": "string"
                },
                "app_impacted": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "date": {
                    "description": "Date is normalised to midnight in Asia/Jakarta.",
                    "type": "string"
                },
                "end_time": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "project_id": {
                    "type": "string"
                },
                "project_name": {
                    "type": "string"
                },
                "project_ref": {
                    "$ref": "#/definitions/models.Project"
                },
                "project_ref_id": {
                    "type": "integer"
                },
                "start_time": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "status_ref": {
                    "$ref": "#/definitions/models.ActivityStatus"
                },
                "updated_at": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Enter JWT token with format \"Bearer {token}\".",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}